moved image drawing to generic image api
